requirements-railway.txt
# Railway专用依赖文件
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==0.28.1
scikit-learn==1.3.0
numpy==1.24.3
gunicorn==21.2.0
